requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-telegram-bot>=20.0
python-dotenv>=0.19.0
aiohttp>=3.11.16

